refactor: separate folder by responsability
