+++ src/modules/alerts/AlertDetailsView.vue
@@ -100,8 +100,8 @@
 import { ref, computed, onMounted, onUnmounted } from 'vue'
 import { useRouter } from 'vue-router'
 import alertServices from '@/modules/alerts/services/alertServices'
-import problemsService from '@/modules/alerts/services/problemsServices'
-import usersService from '@/modules/alerts/services/userServices'
+import problemsService from '@/modules/problems/services/problemsServices'
+import usersService from '@/modules/users/services/userServices'
 import type { AlertDetails } from '@/modules/alerts/types/alertsTypes'
 import type { Problem } from '@/modules/alerts/types/alertsTypes'
 import type { Agent } from '@/modules/alerts/types/alertsTypes'
