+++ src/modules/alerts/mock/routes/alertRoutes.ts
@@ -130,20 +130,23 @@ const alertsRoutes = [
       result: (params: MockParams) => {
         const alertId = Number(params.id)
 
-        const mockAlerts: Record<number, {
-          id: number
-          alert_id: number
-          indicator: string
-          currentLevel: number
-          previousLevel: number
-          location: string
-          timestamp: string
-          finalized: boolean
-          radar_id: string
-          zone: string
-          description: string
-          affected_radars: string[]
-        }> = {
+        const mockAlerts: Record<
+          number,
+          {
+            id: number
+            alert_id: number
+            indicator: string
+            currentLevel: number
+            previousLevel: number
+            location: string
+            timestamp: string
+            finalized: boolean
+            radar_id: string
+            zone: string
+            description: string
+            affected_radars: string[]
+          }
+        > = {
           101: {
             id: 1,
             alert_id: 101,
@@ -206,156 +209,6 @@ const alertsRoutes = [
         return APIFailureWrapper({
           content: response,
           errorMessage: 'Erro ao buscar detalhes do alerta',
-        })
-      },
-    },
-    'on',
-  ),
-
-  mockFlag(
-    {
-      method: 'get',
-      url: '/problems',
-      result: () => {
-        const response = [
-          { id: 1, name: 'Acidente', description: 'Colisão entre veículos' },
-          { id: 2, name: 'Veículo quebrado', description: 'Veículo parado na via' },
-          { id: 3, name: 'Obra na via', description: 'Manutenção ou construção' },
-          { id: 4, name: 'Manifestação', description: 'Protesto ou evento' },
-          { id: 5, name: 'Alagamento', description: 'Via alagada' },
-          { id: 6, name: 'Semáforo com defeito', description: 'Sinalização não funcionando' },
-        ]
-
-        return APIFailureWrapper({
-          content: { items: response },
-          errorMessage: 'Erro ao listar problemas',
-        })
-      },
-    },
-    'on',
-  ),
-
-  mockFlag(
-    {
-      method: 'get',
-      url: '/protocols/:problemId',
-      result: (params: MockParams) => {
-        const protocols: Record<number, {
-          id: number
-          problem_id: number
-          title: string
-          steps: string[]
-        }> = {
-          1: {
-            id: 1,
-            problem_id: 1,
-            title: 'Protocolo para Acidentes',
-            steps: [
-              'Isolar a área do acidente com cones e faixas de segurança',
-              'Acionar SAMU se houver feridos graves',
-              'Fotografar a cena e coletar informações dos envolvidos',
-              'Acionar guincho para remoção dos veículos',
-              'Registrar ocorrência no sistema com todas as informações coletadas',
-              'Liberar a via após remoção completa dos veículos',
-              'Notificar o gestor sobre a conclusão',
-            ],
-          },
-          2: {
-            id: 2,
-            problem_id: 2,
-            title: 'Protocolo para Veículo Quebrado',
-            steps: [
-              'Contatar o motorista do veículo',
-              'Solicitar que sinalize o veículo com triângulo',
-              'Acionar guincho para remoção',
-              'Orientar sobre estacionamento irregular se aplicável',
-              'Aguardar remoção do veículo',
-              'Confirmar liberação da via',
-            ],
-          },
-          3: {
-            id: 3,
-            problem_id: 3,
-            title: 'Protocolo para Obras na Via',
-            steps: [
-              'Verificar se há autorização para a obra',
-              'Conferir sinalização adequada no local',
-              'Solicitar ajustes na sinalização se necessário',
-              'Registrar no sistema as informações da obra',
-              'Monitorar impacto no tráfego',
-              'Manter contato com responsável pela obra',
-            ],
-          },
-          4: {
-            id: 4,
-            problem_id: 4,
-            title: 'Protocolo para Manifestação',
-            steps: [
-              'Avaliar o tamanho e impacto da manifestação',
-              'Acionar Polícia Militar se necessário',
-              'Estabelecer rotas alternativas',
-              'Comunicar aos órgãos competentes',
-              'Monitorar evolução da situação',
-              'Registrar informações detalhadas no sistema',
-            ],
-          },
-          5: {
-            id: 5,
-            problem_id: 5,
-            title: 'Protocolo para Alagamento',
-            steps: [
-              'Isolar a área alagada',
-              'Acionar Defesa Civil',
-              'Estabelecer rotas alternativas',
-              'Alertar motoristas sobre o perigo',
-              'Acompanhar drenagem da água',
-              'Liberar via somente após segurança confirmada',
-            ],
-          },
-          6: {
-            id: 6,
-            problem_id: 6,
-            title: 'Protocolo para Semáforo com Defeito',
-            steps: [
-              'Identificar o tipo de defeito no semáforo',
-              'Acionar equipe de manutenção',
-              'Avaliar necessidade de agentes de trânsito no local',
-              'Informar aos motoristas sobre o problema',
-              'Aguardar reparo',
-              'Confirmar funcionamento após manutenção',
-            ],
-          },
-        }
-
-        const problemId = Number(params.problemId)
-        const protocol = protocols[problemId]
-
-        return APIFailureWrapper({
-          content: protocol || null,
-          errorMessage: 'Erro ao buscar protocolo',
-        })
-      },
-    },
-    'on',
-  ),
-
-  mockFlag(
-    {
-      method: 'get',
-      url: '/agents',
-      result: () => {
-        const response = [
-          { id: 1, name: 'João Silva', phone: '5512991234567', available: true },
-          { id: 2, name: 'Maria Santos', phone: '5512992345678', available: true },
-          { id: 3, name: 'Pedro Costa', phone: '5512993456789', available: false },
-          { id: 4, name: 'Ana Oliveira', phone: '5512994567890', available: true },
-          { id: 5, name: 'Carlos Souza', phone: '5512995678901', available: true },
-          { id: 6, name: 'Juliana Lima', phone: '5512996789012', available: false },
-        ]
-
-        return APIFailureWrapper({
-          content: { items: response },
-          errorMessage: 'Erro ao listar agentes',
         })
       },
     },
